+++ bogenliga/src/app/modules/verwaltung/components/liga/liga-detail/liga-detail.component.ts
@@ -23,6 +23,9 @@ import {RegionDO} from '../../../types/region-do.class';
 import {LIGA_DETAIL_CONFIG} from './liga-detail.config';
 import {SessionHandling} from '@shared/event-handling';
 import {CurrentUserService, OnOfflineService} from '@shared/services';
+import {DisziplinDO} from '@verwaltung/types/disziplin-do.class';
+import {DisziplinDTO} from '@verwaltung/types/datatransfer/disziplin-dto.class';
+import {DisziplinDataProviderService} from '@verwaltung/services/disziplin-data-provider-service';
 
 const ID_PATH_PARAM = 'id';
 const NOTIFICATION_DELETE_LIGA = 'liga_detail_delete';
@@ -44,6 +47,9 @@ export class LigaDetailComponent extends CommonComponentDirective implements OnI
   public currentUbergeordneteLiga: LigaDO = new LigaDO();
   public allUebergeordnete: Array<LigaDO> = [new LigaDO()];
 
+  public currentDisziplin: DisziplinDO;
+  public allDisziplin: Array<DisziplinDO> = [new DisziplinDO()];
+
   public currentRegion: RegionDO = new RegionDO();
   public regionen: Array<RegionDO> = [new RegionDO()];
 
@@ -59,6 +65,7 @@ export class LigaDetailComponent extends CommonComponentDirective implements OnI
   constructor(private ligaDataProvider: LigaDataProviderService,
     private regionProvider: RegionDataProviderService,
     private userProvider: UserProfileDataProviderService,
+    private disziplinDataProvider: DisziplinDataProviderService,
     private router: Router,
     private route: ActivatedRoute,
     private notificationService: NotificationService,
@@ -77,7 +84,9 @@ export class LigaDetailComponent extends CommonComponentDirective implements OnI
         this.id = params[ID_PATH_PARAM];
         if (this.id === 'add') {
           this.currentLiga = new LigaDO();
+          this.currentDisziplin = new DisziplinDO();
 
+          this.loadDisziplin();
           this.loadUebergeordnete(); // additional Request for all 'liga' to get all uebergeordnete
           this.loadRegions(); // Request all regions from backend
           this.loadUsers();
@@ -246,6 +255,12 @@ export class LigaDetailComponent extends CommonComponentDirective implements OnI
         .catch((response: BogenligaResponse<LigaDO>) => this.handleFailure(response));
   }
 
+  private loadDisziplin() {
+    this.disziplinDataProvider.findAll()
+        .then((response: BogenligaResponse<DisziplinDO[]>) => this.handleDisziplinResponseArraySuccess(response))
+        .catch((response: BogenligaResponse<DisziplinDTO[]>) => this.handleDisziplinResponseArrayFailure(response));
+  }
+
   private loadRegions() {
     this.regionProvider.findAll()
       .then((response: BogenligaResponse<RegionDO[]>) => this.handleResponseArraySuccess(response))
@@ -322,6 +337,18 @@ export class LigaDetailComponent extends CommonComponentDirective implements OnI
         });
 
     this.notificationService.showNotification(notification);
+  }
+
+  private handleDisziplinResponseArraySuccess(response: BogenligaResponse<DisziplinDO[]>): void {
+    this.allDisziplin = [];
+    this.allDisziplin = response.payload;
+    this.loading = false;
+  }
+
+
+  private handleDisziplinResponseArrayFailure(response: BogenligaResponse<DisziplinDTO[]>): void {
+    this.allDisziplin = [];
+    this.loading = false;
   }
 
   private handleResponseArraySuccess(response: BogenligaResponse<RegionDO[]>): void {
